Protocol to try to break up app for cleaner tool shed install typing.
