+++ lib/galaxy/app_unittest_utils/galaxy_mock.py
@@ -93,7 +93,7 @@ class MockApp(di.Container, GalaxyDataTestApp):
     config: "MockAppConfig"
     amqp_type: str
     job_search: Optional[JobSearch] = None
-    toolbox: ToolBox
+    _toolbox: ToolBox
     tool_cache: ToolCache
     install_model: ModelMapping
     watchers: ConfigWatchers
@@ -108,6 +108,7 @@ class MockApp(di.Container, GalaxyDataTestApp):
         super().__init__()
         config = config or MockAppConfig(**kwargs)
         GalaxyDataTestApp.__init__(self, config=config, **kwargs)
+        self.install_model = self.model
         self[BasicSharedApp] = cast(BasicSharedApp, self)
         self[MinimalManagerApp] = cast(MinimalManagerApp, self)
         self[StructuredApp] = cast(StructuredApp, self)
@@ -150,6 +151,10 @@ class MockApp(di.Container, GalaxyDataTestApp):
             return "/mock/url"
 
         self.url_for = url_for
+
+    @property
+    def toolbox(self) -> ToolBox:
+        return self._toolbox
 
     def wait_for_toolbox_reload(self, toolbox):
         # TODO: If the tpm test case passes, does the operation really
