+++ lib/galaxy/queue_worker.py
@@ -184,7 +184,7 @@ def _get_new_toolbox(app, save_integrated_tool_panel=True):
     app.datatypes_registry.load_external_metadata_tool(new_toolbox)
     load_lib_tools(new_toolbox)
     [new_toolbox.register_tool(tool) for tool in new_toolbox.data_manager_tools.values()]
-    app.toolbox = new_toolbox
+    app._toolbox = new_toolbox
     app.toolbox.persist_cache()
