+++ lib/galaxy/structured_app.py
@@ -31,6 +31,7 @@ from galaxy.quota import QuotaAgent
 from galaxy.security.idencoding import IdEncodingHelper
 from galaxy.security.vault import Vault
 from galaxy.tool_shed.cache import ToolShedRepositoryCache
+from galaxy.tool_util.data import ToolDataTableManager
 from galaxy.tool_util.deps.views import DependencyResolversView
 from galaxy.tool_util.verify import test_data
 from galaxy.util.dbkeys import GenomeBuilds
@@ -48,7 +49,6 @@ if TYPE_CHECKING:
     from galaxy.managers.workflows import WorkflowsManager
     from galaxy.tools import ToolBox
     from galaxy.tools.cache import ToolCache
-    from galaxy.tools.data import ToolDataTableManager
     from galaxy.tools.error_reports import ErrorReports
     from galaxy.visualization.genomes import Genomes
 
@@ -67,9 +67,12 @@ class BasicSharedApp(Container):
     model: SharedModelMapping
     security: IdEncodingHelper
     auth_manager: AuthManager
-    toolbox: "ToolBox"
     security_agent: Any
     quota_agent: QuotaAgent
+
+    @property
+    def toolbox(self) -> "ToolBox":
+        ...
 
 
 class MinimalToolApp(Protocol):
@@ -77,7 +80,7 @@ class MinimalToolApp(Protocol):
     config: Any  # 'galaxy.config.BaseAppConfiguration'
     datatypes_registry: Registry
     object_store: ObjectStore
-    tool_data_tables: "ToolDataTableManager"
+    tool_data_tables: ToolDataTableManager
     file_sources: ConfiguredFileSources
     security: IdEncodingHelper
 
@@ -142,8 +145,8 @@ class StructuredApp(MinimalManagerApp):
     webhooks_registry: WebhooksRegistry
     queue_worker: Any  # 'galaxy.queue_worker.GalaxyQueueWorker'
     data_provider_registry: Any  # 'galaxy.visualization.data_providers.registry.DataProviderRegistry'
-    tool_data_tables: "ToolDataTableManager"
-    tool_cache: "ToolCache"
+    tool_data_tables: ToolDataTableManager
+    tool_cache: ToolCache
     tool_shed_registry: ToolShedRegistry
     tool_shed_repository_cache: Optional[ToolShedRepositoryCache]
     watchers: "ConfigWatchers"
