+++ lib/galaxy/webapps/galaxy/api/tool_data.py
@@ -2,7 +2,7 @@ from fastapi import Path
 from fastapi.responses import FileResponse
 
 from galaxy.managers.tool_data import ToolDataManager
-from galaxy.tools.data._schema import (
+from galaxy.tool_util.data._schema import (
     ToolDataDetails,
     ToolDataEntryList,
     ToolDataField,
