+++ mypy.ini
@@ -248,7 +248,7 @@ check_untyped_defs = False
 check_untyped_defs = False
 [mypy-galaxy.tools.expressions.evaluation]
 check_untyped_defs = False
-[mypy-galaxy.tools.data]
+[mypy-galaxy.tool_util.data]
 check_untyped_defs = False
 [mypy-galaxy.tool_util.verify]
 check_untyped_defs = False
